Backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Resume processing and email sending application"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.0.0",
    "reportlab>=4.0.0",
    "streamlit>=1.28.0",
    "groq>=0.4.0",
    "python-dotenv>=0.19.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "dropbox>=11.0.0",
    "sendgrid>=6.0.0",
    "pandas>=1.5.0",
]
